requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
anyio==3.3.3; python_full_version >= '3.6.2'
asgiref==3.4.1; python_version >= '3.6'
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3; python_version >= '3.6'
dnspython==2.1.0; python_version >= '3.6'
email-validator==1.1.3
fastapi==0.70.0
greenlet==1.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
h11==0.12.0; python_version >= '3.6'
idna==3.3; python_version >= '3.5'
passlib==1.7.4
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pydantic[email]==1.8.2
pyjwt==2.2.0
python-multipart==0.0.5
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.2.0; python_version >= '3.5'
sqlalchemy==1.4.25
starlette==0.16.0; python_version >= '3.6'
typing-extensions==3.10.0.2
uvicorn==0.15.0
